# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HTCMC PCHSystem",
    "version": "0.10.0"
  },
  "paths": {
...
    "/sheets": {
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "新建表格",
        "operationId": "create_sheet_sheets_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sheets"
        ],
        "summary": "表格列表（status/owner 过滤）",
        "operationId": "list_sheets_sheets_get",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "过滤：传 me 只看自己",
              "title": "Owner"
            },
            "description": "过滤：传 me 只看自己"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "阶段过滤：collecting / constructing / archived / active",
              "title": "Status"
            },
            "description": "阶段过滤：collecting / constructing / archived / active"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SheetSummary"
                  },
                  "title": "Response List Sheets Sheets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sheets/{sheet_id}": {
      "get": {
        "tags": [
          "sheets"
        ],
        "summary": "表详情（?format=csv 返回 CSV）",
        "operationId": "get_sheet_sheets__sheet_id__get",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "传 csv 返回 text/csv",
              "title": "Format"
            },
            "description": "传 csv 返回 text/csv"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按 item_name/registry_id 大小写不敏感过滤行",
              "title": "Q"
            },
            "description": "按 item_name/registry_id 大小写不敏感过滤行"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sheets"
        ],
        "summary": "更新表（标题等）",
        "operationId": "patch_sheet_sheets__sheet_id__patch",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sheets"
        ],
        "summary": "删除表",
        "operationId": "delete_sheet_sheets__sheet_id__delete",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sheets/{sheet_id}/rows/{row_id}/claim": {
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "认领行",
        "operationId": "claim_row_sheets__sheet_id__rows__row_id__claim_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sheets/{sheet_id}/rows/{row_id}/release": {
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "解除认领",
        "operationId": "release_row_sheets__sheet_id__rows__row_id__release_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sheets/{sheet_id}/rows/{row_id}/reject": {
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "打回行（重开）",
        "operationId": "reject_row_sheets__sheet_id__rows__row_id__reject_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sheets/{sheet_id}/rows/{row_id}/contribute": {
      "post": {
        "tags": [
          "sheets"
        ],
        "summary": "提交贡献数量",
        "description": "progress 行增量上交（任意登录玩家）。",
        "operationId": "contribute_to_row_sheets__sheet_id__rows__row_id__contribute_post",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowContributeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sheets/{sheet_id}/rows/{row_id}/progress": {
      "patch": {
        "tags": [
          "sheets"
        ],
        "summary": "更新进度数量",
        "description": "progress 行 owner 直接修正进度（绝对值，可增可减）。",
        "operationId": "set_row_progress_sheets__sheet_id__rows__row_id__progress_patch",
        "security": [
          {
            "Authorization": []
          },
          {
            "X-Service-Token": []
          },
          {
            "X-Player-UUID": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sheet Id"
            }
          },
          {
            "name": "row_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowProgressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SheetCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SheetCreateRequest"
      },
...
      "SheetSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "owner_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Uuid"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "archived_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived Path"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_uuid",
          "owner_name",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SheetSummary"
      },
...
    },
    "securitySchemes": {
      "X-Service-Token": {
        "type": "apiKey",
        "description": "服务端组件令牌（MCDR / 服务端 mod / 服主脚本）",
        "in": "header",
        "name": "X-Service-Token"
      },
      "Authorization": {
        "type": "apiKey",
        "description": "Web JWT 通道，值形如 \"Bearer <jwt>\"",
        "in": "header",
        "name": "Authorization"
      },
      "X-Player-UUID": {
        "type": "apiKey",
        "description": "MCDR 代理通道玩家 UUID（与 X-Service-Token 成对）",
        "in": "header",
        "name": "X-Player-UUID"
      },
      "X-Source-Id": {
        "type": "apiKey",
        "description": "施工上报源标识（服务端 mod 白名单名）",
        "in": "header",
        "name": "X-Source-Id"
      }
    }
  }
}